requirements.txt
pillow==10.4.0
PyMuPDF==1.24.11
PyYAML==6.0.2
requests==2.32.3
